requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
pyyaml>=6.0
mlflow>=2.9
xlsxwriter>=3.1

# FactSet Formula API (data_extraction.py — optional if you only use regime_detection)
fds.sdk.Formula>=2.0
fds.sdk.utils>=1.0
